test: moved tests from model-client to model-datastructure
