+++ model-datastructure/src/commonTest/kotlin/RandomTreeChangeGenerator.kt
@@ -1,15 +1,29 @@
-package org.modelix.model
+/*
+ * Copyright (c) 2023.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
 import org.modelix.model.api.IBranch
 import org.modelix.model.api.IConcept
+import org.modelix.model.api.IIdGenerator
 import org.modelix.model.api.ITree
 import org.modelix.model.api.IWriteTransaction
 import org.modelix.model.api.PBranch
 import org.modelix.model.api.PNodeReference
-import org.modelix.model.client.IdGenerator
 import kotlin.random.Random
 
-class RandomTreeChangeGenerator(private val idGenerator: IdGenerator, private val rand: Random) {
+class RandomTreeChangeGenerator(private val idGenerator: IIdGenerator, private val rand: Random) {
     val childRoles = listOf("cRole1", "cRole2", "cRole3")
     val propertyRoles = listOf("pRole1", "pRole2", "pRole3")
     val referenceRoles = listOf("rRole1", "rRole2", "rRole3")
