+++ model-datastructure/src/commonTest/kotlin/TreeDiffTest.kt
@@ -1,4 +1,6 @@
 /*
+ * Copyright (c) 2023.
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -11,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.modelix.model
 
 import org.modelix.model.api.ITree
 import org.modelix.model.api.ITreeChangeVisitor
