+++ model-datastructure/src/commonTest/kotlin/UndoTest.kt
@@ -1,5 +1,22 @@
-package org.modelix.model
+/*
+ * Copyright (c) 2023.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
+import org.modelix.model.LinearHistory
+import org.modelix.model.VersionMerger
+import org.modelix.model.api.IIdGenerator
 import org.modelix.model.api.ITree
 import org.modelix.model.api.PBranch
 import org.modelix.model.client.IdGenerator
@@ -107,7 +124,7 @@ class UndoTest {
         }
     }
 
-    fun undo(version: CLVersion, idGenerator: IdGenerator): CLVersion {
+    fun undo(version: CLVersion, idGenerator: IIdGenerator): CLVersion {
         return CLVersion.createRegularVersion(
             id = idGenerator.generate(),
             time = null,
@@ -118,7 +135,7 @@ class UndoTest {
         )
     }
 
-    private fun randomChanges(baseBranch: OTBranch, numChanges: Int, idGenerator: IdGenerator, rand: Random) {
+    private fun randomChanges(baseBranch: OTBranch, numChanges: Int, idGenerator: IIdGenerator, rand: Random) {
         baseBranch.runWrite {
             val changeGenerator = RandomTreeChangeGenerator(idGenerator, rand).growingOperationsOnly()
             for (i in 0 until numChanges) {
@@ -130,7 +147,7 @@ class UndoTest {
     fun createVersion(
         opsAndTree: Pair<List<IAppliedOperation>, ITree>,
         previousVersion: CLVersion?,
-        idGenerator: IdGenerator,
+        idGenerator: IIdGenerator,
         storeCache: IDeserializingKeyValueStore,
     ): CLVersion {
         return CLVersion.createRegularVersion(
